+++ rustdoc-json/tests/test_crate/src/bin/main2.rs
